Redo the logic of pip editable local install
We don't do editable local install in build phase. Instead, we pip
install -e .[extras] before running pytest in run script. This is
because build-time local install is not reliable: the egg-info folder is
created at build time in the host's repo. However, when we run the
docker container, we map a different copy of repo into the container,
there won't be an egg-info folder, making pip problematic.